refactor: move the search bar into search folder
