+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import CommentSection from '@components/CommentSection';
 import HeroSection from '@components/HeroSection';
 import NearbyActivity from '@components/NearbyActivity';
 import RecommendActivity from '@components/RecommendActivity';
-import SearchBar from '@components/SearchBar';
+import SearchBar from '@components/search/SearchBar';
 import TopCategoryMenu from '@components/TopCategoryMenu';
 
 export const runtime = 'edge';
