+++ src/components/search/SearchBar/AdvancedSearchBarDesktop.tsx
@@ -5,7 +5,6 @@ import { Button } from '@components/ui/button';
 import cn from '@lib/utils';
 import { XSquare } from 'lucide-react';
 import { MouseEventHandler, useState } from 'react';
-import { SearchField, useSearch } from './SearchProvider';
 import ActivityField, {
   ActivityKeyword,
   ActivityPicture,
@@ -23,6 +22,7 @@ import {
   distances,
   locations,
 } from './fields/utils';
+import { SearchField, useSearch } from './SearchProvider';
 
 export type AdvancedSearchBarDesktopProps = {
   onSearchSubmit?: (value: SearchParams) => void;
