+++ src/components/search/SearchBar/fields/ActivityResultThumbnail.tsx
