+++ src/components/search/SearchBar/fields/utils.tsx
