+++ src/components/search/SearchBar/index.ts
@@ -1,6 +1,6 @@
+import { updateQueryString } from './fields/utils';
 import SearchBar from './SearchBar';
 import SearchPagination, { SearchPaginationProps } from './SearchPagination';
-import { updateQueryString } from './fields/utils';
 
 export { SearchPagination, updateQueryString };
 export type { SearchPaginationProps };
